requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data handling
datasets>=2.16.0
pandas>=2.1.0
numpy>=1.26.0

# RL and ML
torch>=2.1.0

# Evaluation and metrics
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Logging and utilities
tqdm>=4.66.0
loguru>=0.7.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# API rate limiting
ratelimit>=2.2.0
tenacity>=8.2.0
